Fixes issue #8: file not found unless absolute path is given
The call to `JSONSchema.loadfromfile()` did not include the path to the
file, but only the filename itself. This fails if Sphinx is configured
to use separate source and build directories. (note that the check with
`os.access()` uses the absolute path by concatenating the absolute path
with the filename. THis is why the test for the existence of the file
*does not* fail, but the attempt to open the file *does fail*.

This patch uses the same syntax as in the check with `os.access()`, by
concatenating the absolute path to the JSON filename.